(refactor, chore): dpll impl names and packages, add solveAll in Solver
