+++ src/main/scala/satify/update/solver/dpll/impl/DpllAllSol.scala
@@ -1,25 +1,28 @@
-package satify.update.solver.dpll
+package satify.update.solver.dpll.impl
 
-import satify.model.cnf.Bool.False
+import satify.model.Result.*
 import satify.model.cnf.CNF
-import satify.model.cnf.CNF.Symbol
 import satify.model.dpll.DecisionTree.{Branch, Leaf}
-import satify.model.dpll.{Constraint, Decision, DecisionTree, OptionalVariable}
+import satify.model.dpll.PartialAssignment.{extractParAssignmentFromCnf, extractParAssignments}
+import satify.model.dpll.{Decision, DecisionTree}
+import satify.model.{Assignment, Solution}
 import satify.update.solver.dpll.DpllDecision.decide
-import satify.update.solver.dpll.cnf.CNFSimplification.simplifyCnf
 import satify.update.solver.dpll.cnf.CNFSat.{isSat, isUnsat}
-import satify.model.dpll.PartialAssignment.extractParAssignmentFromCnf
 
 import scala.annotation.tailrec
 
-object Dpll:
+object DpllAllSol:
 
-  def dpll(cnf: CNF): DecisionTree = dpll(Decision(extractParAssignmentFromCnf(cnf), cnf))
+  def dpll(cnf: CNF): Solution =
+    val assignments: List[Assignment] =
+      (for partialAssignment <- extractParAssignments(dpll(Decision(extractParAssignmentFromCnf(cnf), cnf)))
+        yield partialAssignment.toAssignments).flatten
+    Solution(if assignments.nonEmpty then SAT else UNSAT, assignments)
 
   /** Main DPLL algorithm.
-    * @param dec first decision
-    * @return decision tree of the run.
-    */
+   * @param dec first decision
+   * @return decision tree of the run.
+   */
   def dpll(dec: Decision): DecisionTree =
 
     case class Frame(d: Decision, done: List[DecisionTree], todos: List[Decision])
