+++ src/test/scala/satify/update/solver/dpll/impl/DpllOneSolTest.scala
@@ -1,17 +1,16 @@
-package satify.update.solver.dpll
+package satify.update.solver.dpll.impl
 
 import org.scalatest.flatspec.AnyFlatSpec
 import org.scalatest.matchers.should.Matchers
-import satify.model.{Assignment, Result, Solution, Variable}
-import satify.model.cnf.Bool.{False, True}
-import satify.model.cnf.CNF.{And, Not, Or, Symbol}
 import satify.model.Result.*
-import satify.model.dpll.{Decision, DecisionTree}
-import satify.model.dpll.DecisionTree.{Branch, Leaf}
+import satify.model.cnf.Bool.{False, True}
 import satify.model.cnf.CNF
-import satify.update.solver.dpll.DpllOneSol.{dpll, resume}
+import satify.model.cnf.CNF.{And, Not, Or, Symbol}
+import satify.model.dpll.DecisionTree.{Branch, Leaf}
 import satify.model.dpll.PartialAssignment.extractParAssignmentFromCnf
-import satify.update.solver.dpll.utils.DpllUtils.extractSolutions
+import satify.model.dpll.{Decision, DecisionTree}
+import satify.model.{Assignment, Result, Solution, Variable}
+import satify.update.solver.dpll.impl.DpllOneSol.{dpll, resume}
 
 class DpllOneSolTest extends AnyFlatSpec with Matchers:
